water_ai 3/requirements.txt
flask
numpy
pillow
scikit-learn
joblib
